[ops] moves server files to proper folder
