+++ bin/entry.js
@@ -12,4 +12,4 @@ if (environment === 'production') {
   process.env.NODE_ENV = 'production';
 }
 
-require('../index');
+require('../src/index');
